zpay32: migrate fuzz tests
